// sourcecode/BTrak.Packages/snova-hrmangement/module/module-components/src/lib/hrmanagment/models/employee-dependent-contact-model.ts
export class EmployeeDependentContactModel {
    employeeDependentId: string;
    employeeId: string;
    firstName: string;
    surName: string;
    userName: string;
    email: string;
    relationshipId: string;
    relationship: string;
    lastName: string;
    otherRelation: string;
    homeTelephone: string;
    mobileNo: string;
    workTelephone: string;
    isEmergencyContact: boolean;
    isDependentContact: boolean;
    stateOrProvinceId: string;
    stateName: string;
    zipOrPostalCode: string;
    addressStreetOne: string;
    addressStreetTwo: string;
    countryId: string;
    countryName: string;
    timeStamp: any;
    totalCount: number;
    isArchived: boolean;
    emergencyContactId: string;
}